+++ src/ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: mamichal <mamichal@student.42warsaw.pl>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/01 19:42:07 by mamichal          #+#    #+#             */
-/*   Updated: 2024/02/27 17:28:55 by mamichal         ###   ########.fr       */
+/*   Updated: 2024/03/06 12:10:41 by mamichal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@
  */
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	int		i;
+	size_t	i;
 	char	*new;
 
 	i = 0;
